modularize wpdetect/__main__.py - update bump_version.py to update version in wpdetect/constants.py - update bump_version location in release-tag.sh
