+++ wpdetect/cli_store.py
@@ -1,0 +1,10 @@
+"""
+    Contains the cli options store.
+"""
+
+# cli options store
+cli_options = {
+    'verbose': True,  # default value
+    'show_signature': False,  # default value
+    'scan_full': False,  # default value
+}
