+++ wpdetect/constants.py
@@ -1,0 +1,12 @@
+"""
+    Contains all the constants used in wpdetect.
+"""
+
+HEADER = "'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_9_3) AppleWebKit/537.75.14" + \
+    " (KHTML, like Gecko) Version/7.0.3 Safari/7046A194A'"
+VERSION = "1.4.9"
+
+# Error messages
+ERROR_UNABLE_TO_OPEN_URL = "Couldn't open url," + \
+    " please make sure to type a valid and publicly accessible url.\n"
+ERROR_UNABLE_TO_OPEN_FILE = "Please enter the file name correctly, file not found!\n"
